requirements.txt
numpy==1.20.3
pandas-bokeh==0.5.5
plotly==4.14.3
streamlit==1.3.1
ta==0.8.0
TA-Lib==1.20.3
yfinance==0.1.67
